Add TraceableBaseClass
